+++ .github/workflows/tests.yml
@@ -3,11 +3,14 @@ on: push
 
 jobs:
     black:
-        name: Black
+        name: gofmt
         runs-on: ubuntu-22.04
         steps:
             - uses: actions/checkout@v2
-            - uses: jpetrucciani/black-check@master
+            - uses: actions/setup-go@v5
+              with:
+                  go-version: 1.24.4
+            - run: cd scraper && gofmt -w .
     prettier:
         name: Prettier
         runs-on: ubuntu-22.04
